Move v4 metadata models to ecs-agent module (#3715)
